move createpresentation method to tests in order to phase it out
